requirements.txt
fastapi
uvicorn
pymongo
passlib[bcrypt]
python-jose[cryptography]
email-validator